Refactor for dependency isolation
