+++ python/cudaq/operator/helpers.py
@@ -9,6 +9,7 @@
 import inspect, itertools, numpy, os, re, sys, typing  # type: ignore
 from typing import Any, Callable, Iterable, Mapping, Optional, Sequence, Tuple
 from numpy.typing import NDArray
+from enum import Enum
 
 from ..mlir._mlir_libs._quakeDialects import cudaq_runtime
 
@@ -20,6 +21,12 @@ elif (3, 10) <= sys.version_info:
 else:
     NumericType = typing.Union[numpy.complexfloating, complex, float, int]
 
+class InitialState(Enum):
+    """
+    Enum to specify the initial quantum state.
+    """
+    ZERO = 1
+    UNIFORM = 2
 
 class _OperatorHelpers:
